web/requirements.txt
python-fasthtml>=0.12.12
python-dotenv>=1.0.0
httpx==0.27.0
uvicorn[standard]>=0.22.0